+++ packages/daf-trust-graph/src/index.ts
@@ -1,4 +1,4 @@
 export { ServiceController } from './service-controller'
 export { ActionHandler, ActionSendJWT, ActionTypes } from './action-handler'
-import * as Gql from './graphql'
+import Gql from './graphql'
 export { Gql }