+++ CHANGELOG.md
@@ -3,6 +3,82 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.17](https://github.com/uport-project/daf/compare/v2.3.16...v2.3.17) (2020-03-03)
+
+
+### Bug Fixes
+
+* Ethr-did toEthereumAddress ([7335390](https://github.com/uport-project/daf/commit/733539014eb4104bb8a4f82b42fef92cc7f0c067))
+* VP aud ([bc9f498](https://github.com/uport-project/daf/commit/bc9f4988c66fbeadb7e6ce38aec446fd5d38339d))
+
+
+
+
+
+## [2.3.16](https://github.com/uport-project/daf/compare/v2.3.15...v2.3.16) (2020-03-03)
+
+**Note:** Version bump only for package daf
+
+
+
+
+
+## [2.3.15](https://github.com/uport-project/daf/compare/v2.3.14...v2.3.15) (2020-03-02)
+
+
+### Bug Fixes
+
+* GQL export ([63b9c23](https://github.com/uport-project/daf/commit/63b9c23e21e297347e8b24e1ed0cfdb6bd8d5524))
+* GQL export ([eacc969](https://github.com/uport-project/daf/commit/eacc9697e87bbf6a91db8d9bd5195aea096a13c4))
+* Resolver config ([2fdba37](https://github.com/uport-project/daf/commit/2fdba37833c5ad7f40a9bdb70e69eab91ce291ca))
+* Typescript types ([72c1899](https://github.com/uport-project/daf/commit/72c18993ddba6a7a75ae8397e6549cdd29dccb31))
+
+
+
+
+
+## [2.3.14](https://github.com/uport-project/daf/compare/v2.3.13...v2.3.14) (2020-03-02)
+
+
+### Bug Fixes
+
+* TG findEdges fromDID ([dc63238](https://github.com/uport-project/daf/commit/dc63238b34d8d448ded854c7635a55750d02f122))
+
+
+
+
+
+## [2.3.13](https://github.com/uport-project/daf/compare/v2.3.12...v2.3.13) (2020-02-26)
+
+
+### Bug Fixes
+
+* Ethr-did identity provider web3Provider ([a5d6af2](https://github.com/uport-project/daf/commit/a5d6af2032130a7eff1722bc47b758a824906be6))
+* Remove unused ws ([55e0fcb](https://github.com/uport-project/daf/commit/55e0fcb1a68e86c7003909fb03f92ee1b357026c)), closes [#91](https://github.com/uport-project/daf/issues/91)
+
+
+
+
+
+## [2.3.12](https://github.com/uport-project/daf/compare/v2.3.11...v2.3.12) (2020-02-25)
+
+**Note:** Version bump only for package daf
+
+
+
+
+
+## [2.3.11](https://github.com/uport-project/daf/compare/v2.3.10...v2.3.11) (2020-02-25)
+
+
+### Bug Fixes
+
+* Data store await for meta data ([5d96401](https://github.com/uport-project/daf/commit/5d96401530025b76007669e7d28a394378a4fbe9))
+
+
+
+
+
 ## [2.3.10](https://github.com/uport-project/daf/compare/v2.3.9...v2.3.10) (2020-02-25)